backend/requirements.txt
Flask==0.12.2
Flask-Cors==3.0.2
http://download.pytorch.org/whl/cu80/torch-0.3.0.post4-cp35-cp35m-linux_x86_64.whl
torchvision
